Work on project. Stage 1/4: Beginning
Project: Credit Calculator

Description
Let's think about what credit calculator should be able to do. In general, it takes several parameters like a credit principal and a credit interest, calculates the needed values (for example, monthly payment or overpayment) and outputs them.

Not familiar with these concepts? Don't worry, we will explain them to you in simple terms. The principal is the original amount of money you get on credit. And the interest is a charge for borrowed money, usually calculated as a percentage of such a loan.

Objective
Let's start by imitating this behavior. There are some prepared variables in source code that are ready for use: these are text messages that our credit calculator could output. At this stage, all you need to do is output them in the right order.