Stage #3: Annuity payment
Description

Let's compute all the parameters of the credit. There are at least two kinds of credit: those with annuity payment and with differentiated payment. At this stage, you're going to calculate only the annuity payment which is fixed during the whole credit term.

This is the formula:

A(ordinary_annuity)= P * (i * (1 + i) ** n) / ((1 + i) ** n − 1)

Where:

A = annuity payment

P = Credit principal.

i  = nominal (monthly) interest rate. Usually, it’s 1/12 of the annual interest rate. And usually, it’s a floating value, not a percentage. For example, if you have annual interest rate = 12%, then i = 0.01.

n= Number of payments. Usually, it’s the count of months.

You are interested in four values: the count of periods to repay the credit, monthly payment, credit principal and credit interest. Each of these values can be calculated if others are known:

Credit principal:

P = A / ((i * (1 + i) ** n) / ((1 + i) ** n − 1))

A number of payments:

n = log(A / (A − i * P), 1 + i)

Objectives

At this stage, you should add new behavior to the calculator:

First, you should ask the user which parameter they want to calculate. The calculator should be able to calculate the count of periods, monthly payment and credit principal.
Then you need to ask them to input the remaining values.
Finally, compute and output the value that they wanted.
Note that users input interest rate as a percentage, for example, 11.7, so you should divide this value by 100 to use in the formula above.
Please be accurate when converting "X months" to "Y years and Z months". Avoid writing "0 years and 11 months" (output "11 months" instead) and "1 years and 0 months" (output "1 year" instead).

Note that at this stage you have to ask the user to input parameters in a determined order which is provided below. Simply disregard the value the user wants to calculate from this order and follow it. For example, it can be monthly payment if the user typed “a” for the question “What do you want to calculate?”. Here is the order:

The first is the credit principal.
The second is a monthly payment.
The next is the count of the period.
The last is the credit interest.